not done yet

+++ Entities/Loan.cs
@@ -9,12 +9,12 @@ namespace GULLEM_NEW_MVC.Entities
         public int Amount { get; set; }
         public int Term { get; set; }
         public string Payment { get; set; } = null!;
-        public double AmountPaid { get; set; }
-        public double InterestAmount { get; set; }
-        public double Total { get; set; }
-        public double Interest { get; set; }
-        public double Deduction { get; set; }
-        public double ReceivableAmount { get; set; }
+        public float AmountPaid { get; set; }
+        public float InterestAmount { get; set; }
+        public float Total { get; set; }
+        public float Interest { get; set; }
+        public float Deduction { get; set; }
+        public float ReceivableAmount { get; set; }
         public string Status { get; set; } = null!;
         public DateTime DueDate { get; set; }
         public byte[] DateCreated { get; set; } = null!;
